+++ packages/isaacscript-common/src/classes/features/other/EdenStartingStatsHealth.ts
@@ -5,8 +5,10 @@ import {
   mapGetPlayer,
   mapSetPlayer,
 } from "../../../functions/playerDataStructures";
+import { getPlayerHealth } from "../../../functions/playerHealth";
 import { isEden } from "../../../functions/players";
 import { getPlayerStats } from "../../../functions/stats";
+import type { PlayerHealth } from "../../../interfaces/PlayerHealth";
 import type { PlayerStats } from "../../../interfaces/PlayerStats";
 import type { PlayerIndex } from "../../../types/PlayerIndex";
 import { Feature } from "../../private/Feature";
@@ -14,10 +16,11 @@ import { Feature } from "../../private/Feature";
 const v = {
   run: {
     edenPlayerStats: new Map<PlayerIndex, PlayerStats>(),
+    edenPlayerHealth: new Map<PlayerIndex, PlayerHealth>(),
   },
 };
 
-export class EdenStartingStats extends Feature {
+export class EdenStartingStatsHealth extends Feature {
   /** @internal */
   public override v = v;
 
@@ -40,6 +43,11 @@ export class EdenStartingStats extends Feature {
       return;
     }
 
+    this.getEdenStats(player);
+    this.getEdenHealth(player);
+  };
+
+  private getEdenStats(player: EntityPlayer): void {
     const existingStatMap = mapGetPlayer(v.run.edenPlayerStats, player);
     if (existingStatMap !== undefined) {
       return;
@@ -47,11 +55,36 @@ export class EdenStartingStats extends Feature {
 
     const playerStats = getPlayerStats(player);
     mapSetPlayer(v.run.edenPlayerStats, player, playerStats);
-  };
+  }
+
+  private getEdenHealth(player: EntityPlayer): void {
+    const existingHealthMap = mapGetPlayer(v.run.edenPlayerHealth, player);
+    if (existingHealthMap !== undefined) {
+      return;
+    }
+
+    const playerHealth = getPlayerHealth(player);
+    mapSetPlayer(v.run.edenPlayerHealth, player, playerHealth);
+  }
 
   /**
-   * Helper function to get the value of the randomized starting stat for Eden. (At the beginning of
-   * a run, Eden starts with randomized stats.)
+   * Helper function to get the health that Eden started with at the beginning of the run before any
+   * of the random collectibles were added.
+   *
+   * Returns undefined if passed a player that is not Eden.
+   *
+   * In order to use this function, you must upgrade your mod with `ISCFeature.EDEN_STARTING_STATS`.
+   */
+  @Exported
+  public getEdenStartingHealth(
+    player: EntityPlayer,
+  ): Readonly<PlayerHealth> | undefined {
+    return mapGetPlayer(v.run.edenPlayerHealth, player);
+  }
+
+  /**
+   * Helper function to get the value of the randomized starting stat for Eden that was assigned at
+   * the beginning of the run before any of the random collectibles were added.
    *
    * Returns undefined if passed a player that is not Eden.
    *
@@ -72,7 +105,7 @@ export class EdenStartingStats extends Feature {
 
   /**
    * Helper function to get all of the stat values that Eden started with at the beginning of the
-   * run.
+   * run before any of the random collectibles were added.
    *
    * Returns undefined if passed a player that is not Eden.
    *