+++ packages/docs/docs/main/change-log.md
@@ -26,6 +26,7 @@ This page lists the changes to the IsaacScript framework.
   - `CHEST_PICKUP_VARIANTS_SET`
 - Renamed the following features:
   - `ISCFeature.PLAYER_INVENTORY` --> `ISCFeature.PLAYER_COLLECTIBLE_TRACKING`
+  - `ISCFeature.EDEN_STARTING_STATS` --> `ISCFeature.EDEN_STARTING_STATS_HEALTH`
 
 ## September 25th, 2023